requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
fire==0.1.3
firefly-python==0.1.15
gunicorn==19.9.0
idna==2.8
joblib==0.13.2
more-itertools==7.0.0
numpy==1.16.3
pandas==0.24.2
pluggy==0.11.0
py==1.8.0
pymongo==3.8.0
pytest==4.5.0
python-dateutil==2.8.0
python-dotenv==0.10.2
pytz==2019.1
PyYAML==5.1
requests==2.22.0
scikit-learn==0.21.1
scipy==1.3.0
six==1.12.0
urllib3==1.25.3
wcwidth==0.1.7
WebOb==1.8.5
